refactor: Move getMapStyle and related logic to utils
